more sliCQ cleanups
git-svn-id: https://svn.grrrr.org/projects/dev/nsgt@854 3d3ad2b0-3bb8-11dd-8da9-e1ab7d56d6ab

+++ nsgt/nsgfwin.py
@@ -77,7 +77,7 @@ def nsgfwin(fmin,fmax,bins,sr,Ls,min_win=4):
         M[k] = N.round(fbas[k+1]-fbas[k-1])
     M[-1] = N.round(Ls-fbas[-2])
     
-    M = N.clip(M,min_win,N.inf)
+    M = N.clip(M,min_win,N.inf).astype(int)
     
     g = [hannwin(m) for m in M]
